+++ python/lsst/ts/IntegrationTests/auxtel_housekeeping.py
@@ -35,10 +35,12 @@ class AuxTelHousekeeping(BaseScript):
 
     index: int = 2
     configs: tuple = (
+        registry["atdome_housekeeping"],
         registry["atmcs_housekeeping"],
         registry["atspectrograph_housekeeping"],
     )
     scripts: list = [
+        ("run_command.py", BaseScript.is_standard),
         ("run_command.py", BaseScript.is_standard),
         ("run_command.py", BaseScript.is_standard),
     ]